install: Add option to toggle killing resque workers

+++ install.sh
@@ -7,7 +7,7 @@ usage() {
 	echo
 	echo "**USAGE**"
 	echo
-	echo "$0 [-q QUEUE] [-s SERVER_USER] [-t TEST_USER] [-d WORKING_DIR] [-T TESTERS]"
+	echo "$0 [-kh] [-q QUEUE] [-s SERVER_USER] [-t TEST_USER] [-d WORKING_DIR] [-T TESTERS]"
 	echo
 	echo "**OPTIONS**"
 	echo
@@ -16,6 +16,7 @@ usage() {
 	echo "-t TEST_USER, --user-test TEST_USER: [optional, defaults to \"atetest\"] The user that runs the student code being tested."
 	echo "-d WORKING_DIR, --dir WORKING_DIR: [optional, defaults to \"../\"] The name of the test server working directory."
 	echo "-T TESTERS, --testers TESTERS: [optional] A comma-separated list of tester names to install together with the test server."
+	echo "-k, --kill: [optional] Kills all running Resque workers."
 	echo "-h, --help: [optional] Prints this help."
 	echo
 	echo "**EXAMPLES**"
@@ -23,12 +24,12 @@ usage() {
 	echo "./install.sh"
 	echo "Installs the MarkUs autotester, using the queue \"ate_tests\", server user \"ateserver\" and test user \"atetest\"."
 	echo
-	echo "./install.sh -q some_queue -s server_user -t test_user -d /some/dir -T uam"
-	echo "Installs the MarkUs autotester, using the queue \"some_queue\", server user \"server_user\", test user \"test_user\", working directory \"/some/dir\", and the \"uam\" tester."
+	echo "./install.sh -k -q some_queue -s server_user -t test_user -d /some/dir -T uam"
+	echo "Kills all running Resque workers, then installs the MarkUs autotester, using the queue \"some_queue\", server user \"server_user\", test user \"test_user\", working directory \"/some/dir\", and the \"uam\" tester."
 }
 
-SHORT=q:s:t:d:T:h
-LONG=queue:,user-server:,user-test:,dir:,testers:,help
+SHORT=q:s:t:d:T:kh
+LONG=queue:,user-server:,user-test:,dir:,testers:,kill,help
 PARSED=`getopt -o ${SHORT} -l ${LONG} -n "$0" -- "$@"`
 eval set -- "${PARSED}"
 
@@ -37,6 +38,7 @@ USERSERVER=ateserver
 USERTEST=atetest
 WORKINGDIR=..
 TESTERS=()
+KILL=false
 while true; do
 	case "$1" in
 		-q | --queue )
@@ -59,6 +61,10 @@ while true; do
 			IFS=',' read -a TESTERS <<< "$2"
 			shift 2
 			;;
+		-k | --kill )
+			KILL=true
+			shift
+			;;
 		-h | --help )
 			usage
 			exit
@@ -80,8 +86,10 @@ TESTSDIR=${WORKINGDIR}/tests
 RESULTSDIR=${WORKINGDIR}/test_runs
 VENVSDIR=${WORKINGDIR}/venvs
 
-echo "[AUTOTEST] Killing running Resque workers"
-kill -QUIT `pgrep -f resque` || { echo "[AUTOTEST] No running Resque worker found, no need to kill them"; }
+if [ "${KILL}" = true ]; then
+	echo "[AUTOTEST] Killing running Resque workers"
+	kill -QUIT `pgrep -f resque` || { echo "[AUTOTEST] No running Resque worker found, no need to kill them"; }
+fi
 cd ${SERVERDIR}
 echo "[AUTOTEST] Installing dependencies"
 bundle install --deployment
